FASE 1: FLOAT->DECIMAL(12,2) + REGLA 18 Docker + fix TypeError Decimal - 10 columnas monetarias migradas a NUMERIC(12,2) en DB - 3 modelos actualizados (pos, catalog, cash) - Fix TypeError float+Decimal en service.py - REGLA 18 documentada: nunca editar archivos con containers corriendo - Bug TypeError Decimal documentado en spec

+++ apps/api/modules/cash/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, Boolean, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, String, Float, Boolean, ForeignKey, DateTime, Numeric
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from core.database import Base
@@ -15,15 +15,15 @@ class CashSession(Base):
     terminal_id = Column(String, nullable=False, index=True)
     employee_id = Column(Integer, ForeignKey("employees.id"), nullable=False)
     employee_name = Column(String, nullable=False)  # Desnormalizado para reportes rápidos
-    opening_float = Column(Float, nullable=False, default=0.0)
+    opening_float = Column(Numeric(12, 2), nullable=False, default=0)
     status = Column(String, nullable=False, default="OPEN")  # OPEN, CLOSED
     opened_at = Column(DateTime, default=datetime.utcnow)
     closed_at = Column(DateTime, nullable=True)
 
     # Montos físicos capturados al cierre
-    physical_cash = Column(Float, nullable=True)
-    physical_credit = Column(Float, nullable=True)
-    physical_debit = Column(Float, nullable=True)
+    physical_cash = Column(Numeric(12, 2), nullable=True)
+    physical_credit = Column(Numeric(12, 2), nullable=True)
+    physical_debit = Column(Numeric(12, 2), nullable=True)
 
     movements = relationship("CashMovement", back_populates="session", cascade="all, delete-orphan")
     tickets = relationship("Ticket", primaryjoin="CashSession.id == Ticket.cash_session_id", back_populates="cash_session")
@@ -39,7 +39,7 @@ class CashMovement(Base):
     id = Column(Integer, primary_key=True, index=True)
     cash_session_id = Column(Integer, ForeignKey("cash_sessions.id"), nullable=False)
     movement_type = Column(String, nullable=False)  # ENTRADA, SALIDA
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(12, 2), nullable=False)
     concept = Column(String, nullable=False)
     created_at = Column(DateTime, default=datetime.utcnow)
